requirements.txt
streamlit
python-dotenv
tavily-python
qdrant-client
sentence-transformers
groq
inngest
beautifulsoup4
requests
